model/requirements.txt
audioread==3.0.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
decorator==5.1.1
filelock==3.14.0
fsspec==2024.6.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
konlpy==0.6.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
lxml==5.2.2
MarkupSafe==2.1.5
mecab-python===0.996-ko-0.9.2
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.59.1
numpy==1.26.4
packaging==24.0
platformdirs==4.2.2
pooch==1.8.2
pycparser==2.22
pydub==0.25.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
soundfile==0.12.1
soxr==0.3.7
sympy==1.12.1
threadpoolctl==3.5.0
sentence-transformers==3.0.0
torch==2.3.1
torchdata==0.7.0a0
torchtext==0.17.0a0
openai-whisper==20231117
typing_extensions==4.12.1
transformers==4.41.2
urllib3==2.2.1
ffmpeg-python
